fix : discountRate 속성 타입 수정
- int -> float로 수정

+++ src/main/java/com/freeder/buclserver/domain/product/dto/ProductDetailDTO.java
@@ -16,7 +16,7 @@ public class ProductDetailDTO {
 	private String brandName;
 	private int salePrice;
 	private int consumerPrice;
-	private int discountRate;
+	private float discountRate;
 	private float averageRating;
 	private LocalDateTime createdAt;
 	private int totalReviewCount;
